+++ src/monkeymaster/CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2020-04-24
+
+### Fixed
+
+- queueJob respects existing query parameters
+
+### Added
+
+- New helpers for query parameters
+- completionCleanupDelayMs option for queueJob
+- nojob option for executeJob
+
+### Deprecated 
+
+- getQueryParameter helper
+- skipCleanupOnCompletion option for queueJob
+
 ## [1.1.0] - 2020-04-19
 
 ### Changed
@@ -24,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - monkeymaster library
 
-[unreleased]: https://github.com/davidtorosyan/command.games/compare/monkeymaster-v1.1.0...HEAD
+[unreleased]: https://github.com/davidtorosyan/command.games/compare/monkeymaster-v1.2.0...HEAD
+[1.2.0]: https://github.com/davidtorosyan/command.games/compare/monkeymaster-v1.1.0...v1.2.0
 [1.1.0]: https://github.com/davidtorosyan/command.games/compare/monkeymaster-v1.0.0...monkeymaster-v1.1.0
 [1.0.0]: https://github.com/davidtorosyan/command.games/releases/tag/monkeymaster-v1.0.0